--- src/pages/Profile/components/PostSection/index.js
@@ -1,3 +1,0 @@
-export { AlbumView } from './AlbumView';
-export { ListView } from './ListView';
-export { PostViewTab } from './PostViewTab';
