+++ src/pages/Profile/components/PostSectionElement/AlbumView.jsx
