+++ src/pages/Profile/components/PostSectionElement/index.js
@@ -1,0 +1,3 @@
+export * from './AlbumView';
+export * from './ListView';
+export * from './PostViewTab';
