+++ src/pages/Profile/components/ProfileSection.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import '../styles/ProfileSection.scss';
-import { UserInfomation, ButtonsYourProfile, ButtonsMyProfile } from './ProfileSection/index';
+import { UserInfomation, ButtonsYourProfile, ButtonsMyProfile } from './ProfileSectionElement';
 
 export function ProfileSection({ data }) {
   const who = localStorage.getItem('user-accountname');
