--- src/pages/Profile/components/ProfileSection/index.js
@@ -1,3 +1,0 @@
-export { UserInfomation } from './UserInformation';
-export { ButtonsYourProfile } from './ButtonsYourProfile';
-export { ButtonsMyProfile } from './ButtonsMyProfile';
