+++ src/pages/Profile/components/ProfileSectionElement/index.js
@@ -1,0 +1,3 @@
+export * from './UserInformation';
+export * from './ButtonsYourProfile';
+export * from './ButtonsMyProfile';
